+++ src/main/java/gdsc/binaryho/imhere/config/JavaMailSenderConfig.java
@@ -8,7 +8,7 @@ import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.JavaMailSenderImpl;
 
 @Configuration
-public class EmailConfig {
+public class JavaMailSenderConfig {
 
     @Value("${mail.smtp.port}")
     private int port;
