+++ src/main/java/gdsc/binaryho/imhere/core/auth/application/AuthService.java
@@ -1,6 +1,8 @@
 package gdsc.binaryho.imhere.core.auth.application;
 
+import gdsc.binaryho.imhere.core.auth.application.port.VerificationCodeRepository;
 import gdsc.binaryho.imhere.core.auth.exception.DuplicateEmailException;
+import gdsc.binaryho.imhere.core.auth.exception.EmailVerificationCodeIncorrectException;
 import gdsc.binaryho.imhere.core.auth.exception.MemberNotFoundException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordFormatMismatchException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordIncorrectException;
@@ -9,6 +11,7 @@ import gdsc.binaryho.imhere.core.auth.model.response.SignInRequestValidationResu
 import gdsc.binaryho.imhere.core.member.Member;
 import gdsc.binaryho.imhere.core.member.Role;
 import gdsc.binaryho.imhere.core.member.infrastructure.MemberRepository;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -21,6 +24,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AuthService {
 
     private final MemberRepository memberRepository;
+    private final VerificationCodeRepository verificationCodeRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     private static final String PASSWORD_REGEX = "^(?=.*[a-zA-Z])(?=.*[0-9])[a-zA-Z0-9]{8,20}$";
@@ -62,6 +66,13 @@ public class AuthService {
     private void validateMatchesPassword(String rawPassword, String encodedPassword) {
         if (!bCryptPasswordEncoder.matches(rawPassword, encodedPassword)) {
             throw PasswordIncorrectException.EXCEPTION;
+        }
+    }
+
+    public void verifyCode(String email, String verificationCode) {
+        String emailVerificationCode = verificationCodeRepository.getByEmail(email);
+        if (!Objects.equals(emailVerificationCode, verificationCode)) {
+            throw EmailVerificationCodeIncorrectException.EXCEPTION;
         }
     }
 }
