+++ src/test/java/gdsc/binaryho/imhere/core/auth/AuthServiceTest.java
@@ -1,12 +1,15 @@
 package gdsc.binaryho.imhere.core.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import gdsc.binaryho.imhere.core.auth.application.AuthService;
+import gdsc.binaryho.imhere.core.auth.application.port.VerificationCodeRepository;
 import gdsc.binaryho.imhere.core.auth.exception.DuplicateEmailException;
+import gdsc.binaryho.imhere.core.auth.exception.EmailVerificationCodeIncorrectException;
 import gdsc.binaryho.imhere.core.auth.exception.MemberNotFoundException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordFormatMismatchException;
 import gdsc.binaryho.imhere.core.auth.exception.PasswordIncorrectException;
@@ -14,7 +17,9 @@ import gdsc.binaryho.imhere.core.auth.model.request.SignInRequest;
 import gdsc.binaryho.imhere.core.auth.model.response.SignInRequestValidationResult;
 import gdsc.binaryho.imhere.core.member.Member;
 import gdsc.binaryho.imhere.core.member.infrastructure.MemberRepository;
+import gdsc.binaryho.imhere.mock.FakeVerificationCodeRepository;
 import javax.transaction.Transactional;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
@@ -26,16 +31,23 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 class AuthServiceTest {
 
     @Autowired
-    private AuthService authService;
-    @Autowired
     private MemberRepository memberRepository;
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
+    private AuthService authService;
 
-    String UNIV_ID = "UNIV_ID";
-    String NAME = "이진호";
-    String PASSWORD = "abcd1234";
-    String DEFAULT_MEMBER_ROLE = "ROLE_STUDENT";
+    VerificationCodeRepository verificationCodeRepository = new FakeVerificationCodeRepository();
+
+    private static final String UNIV_ID = "UNIV_ID";
+    private static final String NAME = "이진호";
+    private static final String PASSWORD = "abcd1234";
+    private static final String DEFAULT_MEMBER_ROLE = "ROLE_STUDENT";
+    private static final String EMAIL = "dlwlsgh4687@gmail.com";
+
+    @BeforeEach
+    void initAuthService() {
+        authService = new AuthService(memberRepository, verificationCodeRepository, bCryptPasswordEncoder);
+    }
 
     @Test
     @Transactional
@@ -106,5 +118,31 @@ class AuthServiceTest {
             authService.validateSignInRequest(new SignInRequest(UNIV_ID, PASSWORD));
 
         assertThat(signInRequestValidationResult.getRoleKey()).isEqualTo(DEFAULT_MEMBER_ROLE);
+    }
+
+    @Test
+    void 인증_코드를_인증할_수_있다() {
+        // given
+        String verificationCode = "imhere forever";
+        verificationCodeRepository.saveWithEmailAsKey(EMAIL, verificationCode);
+
+        // when
+        // then
+        assertThatCode(
+            () -> authService.verifyCode(EMAIL, verificationCode)
+        ).doesNotThrowAnyException();
+    }
+
+    @Test
+    void 인증_코드가_틀린_경우_예외를_발생시킨다() {
+        // given
+        String verificationCode = "imhere forever";
+        verificationCodeRepository.saveWithEmailAsKey(EMAIL, verificationCode);
+
+        // when
+        // then
+        assertThatThrownBy(
+            () -> authService.verifyCode(EMAIL, verificationCode + "wrong code")
+        ).isInstanceOf(EmailVerificationCodeIncorrectException.class);
     }
 }
